rename TaskSearchDev to TaskSearch
